Move installDependenciesAsync to init
- move js-yaml to expo-cli
