+++ packages/expo-cli/package.json
@@ -47,6 +47,7 @@
     "@types/command-exists": "^1.2.0",
     "@types/dateformat": "^3.0.0",
     "@types/fs-extra": "^9.0.1",
+    "@types/js-yaml": "^3.12.2",
     "@types/hapi__joi": "^17.1.2",
     "@types/inquirer": "6.0.3",
     "@types/klaw-sync": "^6.0.0",
@@ -71,6 +72,7 @@
     "@expo/plist": "0.0.9",
     "@expo/results": "^1.0.0",
     "@expo/simple-spinner": "1.0.2",
+    "js-yaml": "^3.13.1",
     "@expo/spawn-async": "1.5.0",
     "@expo/xdl": "57.9.35",
     "@hapi/joi": "^17.1.1",
