+++ packages/expo-cli/src/commands/init.ts
@@ -6,12 +6,14 @@ import chalk from 'chalk';
 import program, { Command } from 'commander';
 import fs from 'fs';
 import getenv from 'getenv';
+import yaml from 'js-yaml';
 import padEnd from 'lodash/padEnd';
 import trimStart from 'lodash/trimStart';
 import npmPackageArg from 'npm-package-arg';
 import ora from 'ora';
 import pacote from 'pacote';
 import path from 'path';
+import semver from 'semver';
 import terminalLink from 'terminal-link';
 import wordwrap from 'wordwrap';
 
@@ -202,7 +204,7 @@ async function action(projectDir: string, command: Command) {
   if (options.install) {
     const installJsDepsStep = logNewSection('Installing JavaScript dependencies.');
     try {
-      await Exp.installDependenciesAsync(projectPath, packageManager, { silent: true });
+      await installDependenciesAsync(projectPath, packageManager, { silent: true });
       installJsDepsStep.succeed('Installed JavaScript dependencies.');
     } catch {
       installJsDepsStep.fail(
@@ -277,6 +279,41 @@ async function action(projectDir: string, command: Command) {
     await initGitRepoAsync(projectPath, { silent: true, commit: true });
   } catch {
     // todo: check if git is installed, bail out
+  }
+}
+
+export async function installDependenciesAsync(
+  projectRoot: string,
+  packageManager: 'yarn' | 'npm',
+  flags: { silent: boolean } = { silent: false }
+) {
+  const options = { cwd: projectRoot, silent: flags.silent };
+  if (packageManager === 'yarn') {
+    const yarn = new PackageManager.YarnPackageManager(options);
+    const version = await yarn.versionAsync();
+    const nodeLinker = await yarn.getConfigAsync('nodeLinker');
+    if (semver.satisfies(version, '>=2.0.0-rc.24') && nodeLinker !== 'node-modules') {
+      const yarnRc = path.join(projectRoot, '.yarnrc.yml');
+      let yamlString = '';
+      try {
+        yamlString = fs.readFileSync(yarnRc, 'utf8');
+      } catch (error) {
+        if (error.code !== 'ENOENT') {
+          throw error;
+        }
+      }
+      const config = yamlString ? yaml.safeLoad(yamlString) : {};
+      config.nodeLinker = 'node-modules';
+      !flags.silent &&
+        log.warn(
+          `Yarn v${version} detected, enabling experimental Yarn v2 support using the node-modules plugin.`
+        );
+      !flags.silent && log(`Writing ${yarnRc}...`);
+      fs.writeFileSync(yarnRc, yaml.safeDump(config));
+    }
+    await yarn.installAsync();
+  } else {
+    await new PackageManager.NpmPackageManager(options).installAsync();
   }
 }
