+++ packages/xdl/package.json
@@ -66,7 +66,6 @@
     "internal-ip": "4.3.0",
     "invariant": "2.2.4",
     "is-reachable": "^4.0.0",
-    "js-yaml": "^3.13.1",
     "latest-version": "5.1.0",
     "lodash": "4.17.15",
     "md5hex": "1.0.0",
@@ -118,7 +117,6 @@
     "@types/getenv": "^0.7.0",
     "@types/hapi__joi": "^17.1.2",
     "@types/hashids": "^1.0.30",
-    "@types/js-yaml": "^3.12.2",
     "@types/minipass": "^2.2.0",
     "@types/ncp": "^2.0.1",
     "@types/node-forge": "^0.8.6",
